FOGi/requirements.txt
flask==3.1.2
flask-socketio==5.5.1
eventlet==0.40.3
python-dotenv==1.2.1
openai==2.7.1
